fix: move generateFeed function to helper module
- Moved the generateFeed function from the atom route to the helper module for better code organization and reusability.
- Updated the atom and rss routes to import the generateFeed function from the new location.
- This change enhances maintainability and streamlines the feed generation process.
